commit IdelateProducto

+++ src/main/java/com/sayuri/wifi/models/Modem.java
@@ -4,7 +4,7 @@ public class Modem extends Producto{
     private int velocidad;
     private String tipoModem;
 
-    public Modem(String nombre, double costo, int velocidad, String tipoModem) {
+    public Modem(String nombre, int costo, int velocidad, String tipoModem) {
         super(nombre, costo);
         this.velocidad = velocidad;
         this.tipoModem = tipoModem;
